# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ambatuwork API v1",
        "description": "API documentation for Ambatuwork, a Scrum-based collaboration application used by Web and Android clients. Authenticate your requests using a Sanctum Bearer token.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 Endpoint Base"
        }
    ],
    "paths": {
...
        "/projects/{project}/sprints/{sprint}/retrospective": {
            "get": {
                "tags": [
                    "Retrospective"
                ],
                "summary": "Get Retrospective details",
                "description": "Returns the retrospective details and feedback items for the specified sprint.",
                "operationId": "d521bba6b86ea58a8e1e0b740f476de2",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Retrospective"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project, Sprint or Retrospective not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Retrospective"
                ],
                "summary": "Submit Retrospective Score",
                "description": "Submits or updates the overall team happiness score for a sprint retrospective.",
                "operationId": "5d04b436f112219008d7f9cdc98d4f60",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "team_happiness_score"
                                ],
                                "properties": {
                                    "team_happiness_score": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retrospective score saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Retrospective"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project or Sprint not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{project}/sprints/{sprint}/retrospective/items": {
            "post": {
                "tags": [
                    "Retrospective"
                ],
                "summary": "Create Retrospective Feedback Item",
                "description": "Creates a feedback item (went well, to improve, or action item) for a sprint retrospective.",
                "operationId": "0a507c908959242c237b9861918ec1c4",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "body"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "went_well",
                                        "enum": [
                                            "went_well",
                                            "to_improve",
                                            "action_item"
                                        ]
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Collaboration was awesome!"
                                    },
                                    "assigned_to_user_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Feedback item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RetroItem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project or Sprint not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{project}/sprints/{sprint}/retrospective/items/{retroItem}": {
            "delete": {
                "tags": [
                    "Retrospective"
                ],
                "summary": "Delete Retrospective Feedback Item",
                "description": "Deletes a feedback item. Users can only delete their own items.",
                "operationId": "874a14bb4600ed3a254a53a9b49a0a87",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "retroItem",
                        "in": "path",
                        "description": "Retrospective Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Retrospective item deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project, Sprint or Item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Retrospective"
                ],
                "summary": "Update Retrospective Feedback Item",
                "description": "Updates a feedback item's body or details. Users can only update their own items.",
                "operationId": "cdec713f01f6312e7c4d8b153d89aa31",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "retroItem",
                        "in": "path",
                        "description": "Retrospective Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "went_well",
                                        "enum": [
                                            "went_well",
                                            "to_improve",
                                            "action_item"
                                        ]
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Collaboration was extremely awesome!"
                                    },
                                    "assigned_to_user_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RetroItem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project, Sprint or Item not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "avatar_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://lh3.googleusercontent.com/...",
                        "nullable": true
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-01T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "RetroItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "went_well"
                    },
                    "body": {
                        "type": "string",
                        "example": "Collaboration was awesome!"
                    },
                    "author_user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Retrospective": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sprint_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "team_happiness_score": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RetroItem"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token. Example: Bearer 1|abcdef...",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}